Begin diagnosis-detail tests
